+++ commands/ConfigCommand.js
@@ -106,6 +106,9 @@ ConfigCommand.prototype = extend(BaseCommand.prototype, {
 	identifyServer: function () {
 		console.log('Current profile: ' + settings.profile);
 		console.log('Using API: ' + settings.apiUrl);
+		if (settings.proxyUrl) {
+			console.log('Proxy URL: ' + settings.proxyUrl);
+		}
 		console.log('Access token: ' + settings.access_token);
 	},